Admin validations added

+++ src/main/java/season11/kino_arena/controller/CinemaController.java
@@ -2,12 +2,15 @@ package season11.kino_arena.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import season11.kino_arena.exceptions.AuthorizationException;
 import season11.kino_arena.exceptions.BadRequestException;
 import season11.kino_arena.exceptions.NotFoundException;
 import season11.kino_arena.model.dao.CinemaDAO;
 import season11.kino_arena.model.dto.MessageDTO;
 import season11.kino_arena.model.pojo.Cinema;
+import season11.kino_arena.model.pojo.User;
 
+import javax.servlet.http.HttpSession;
 import java.sql.SQLException;
 import java.util.ArrayList;
 
@@ -18,21 +21,33 @@ public class CinemaController {
     CinemaDAO cinemaDAO;
 
     @PostMapping("/cinemas/add")
-    public Cinema addCinema(@RequestBody Cinema cinema) throws SQLException, BadRequestException {
+    public Cinema addCinema(@RequestBody Cinema cinema, HttpSession session) throws SQLException, BadRequestException {
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null || !user.getIsAdmin()){
+            throw new AuthorizationException("You don`t have permissions for that");
+        }
         validateCinemaData(cinema);
         cinemaDAO.addCinema(cinema);
         return cinema;
     }
 
     @PutMapping("cinemas")
-    public Cinema editCinema(@RequestBody Cinema cinema) throws SQLException, BadRequestException {
+    public Cinema editCinema(@RequestBody Cinema cinema, HttpSession session) throws SQLException, BadRequestException {
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null || !user.getIsAdmin()){
+            throw new AuthorizationException("You don`t have permissions for that");
+        }
         validateCinemaData(cinema);
         cinemaDAO.updateCinema(cinema);
         return cinema;
     }
 
     @DeleteMapping("cinemas/{id}")
-    public MessageDTO deleteCinema(@PathVariable(name = "id") long id) throws NotFoundException, SQLException {
+    public MessageDTO deleteCinema(@PathVariable(name = "id") long id, HttpSession session) throws NotFoundException, SQLException {
+        User user = (User) session.getAttribute(UserController.SESSION_KEY_LOGGED_USER);
+        if(user == null || !user.getIsAdmin()){
+            throw new AuthorizationException("You don`t have permissions for that");
+        }
         cinemaDAO.deleteCinema(id);
         return new MessageDTO("Cinema deleted successfully.");
     }
